+++ src/main/java/pl/ilonaptak/OptimalizeANDget_DNA/reaction/ReactionController.java
@@ -1,6 +1,7 @@
 package pl.ilonaptak.OptimalizeANDget_DNA.reaction;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -8,6 +9,8 @@ import org.springframework.web.bind.annotation.*;
 import pl.ilonaptak.OptimalizeANDget_DNA.accessory.Accessory;
 import pl.ilonaptak.OptimalizeANDget_DNA.experiment.ExperimentService;
 import pl.ilonaptak.OptimalizeANDget_DNA.ingredient.Ingredient;
+import pl.ilonaptak.OptimalizeANDget_DNA.user.CurrentUser;
+import pl.ilonaptak.OptimalizeANDget_DNA.user.User;
 
 import javax.validation.Valid;
 
@@ -33,7 +36,7 @@ public class ReactionController {
                 model.addAttribute("ingredient", new Ingredient());
                 model.addAttribute("accessory", new Accessory());
             }
-            return "experiment/details";
+            return "redirect:/experiment/details/" + id;
         }
         try {
 //            int experimentId = (int) model.getAttribute("experimentId");
@@ -42,7 +45,7 @@ public class ReactionController {
         } catch (NullPointerException e) {
             e.printStackTrace();
         }
-        return "redirect:ingredient/all"; // TODO zmienić lądowanie dla akcji
+        return "redirect:/experiment/details/" + id;
     }
 
 
@@ -52,29 +55,35 @@ public class ReactionController {
         return "reaction/all";
     }
 
-
-    @GetMapping("/update/{id}")
-    public String update(@PathVariable int id, Model model) {
-        model.addAttribute("reaction", reactionService.getById(id));
-        return "reaction/form";
-    }
+// TODO wymyślić jak ogarnąć update
 
 
-    @PostMapping("/update/{id}")
-    @ResponseBody
-    public String update(@PathVariable int id, @Valid Reaction reaction, BindingResult bindingResult) {
-        if (bindingResult.hasErrors()) {
-            return "reaction/form";
-        }
-        reactionService.save(reaction);
-        return "redirect:user/experiment/all";
-    }
+//    @GetMapping("/update/{id}")
+//    public String update(@PathVariable int id, Model model) {
+//        model.addAttribute("reaction", reactionService.getById(id));
+//        return "reaction/form";
+//    }
+//
+//
+//    @PostMapping("/update/{id}")
+//    @ResponseBody
+//    public String update(@PathVariable int id, @Valid Reaction reaction, BindingResult bindingResult) {
+//        if (bindingResult.hasErrors()) {
+//            return "reaction/form";
+//        }
+//        reactionService.save(reaction);
+//        return "redirect:user/experiment/all";
+//    }
 
     @RequestMapping("/delete/{id}")
     @ResponseBody
-    public String delete(@PathVariable int id) {
-        reactionService.deleteById(id);
-        return "deleted";
+    public String delete(@PathVariable int id, @AuthenticationPrincipal CurrentUser currentUser) {
+        User user = currentUser.getUser();
+        User userExperiment = reactionService.getById(id).getExperiment().getUser();
+        if (user.getId() == userExperiment.getId()) {
+            reactionService.deleteById(id);
+        }
+        return "redirect:/user/account/"+user.getId();
     }
